Move ISettingFetcher to a sub-package
