+++ server/collector/src/main/java/com/sbss/bithon/server/collector/brpc/BrpcCollectorStarter.java
@@ -16,7 +16,7 @@
 
 package com.sbss.bithon.server.collector.brpc;
 
-import com.sbss.bithon.agent.rpc.brpc.ISettingFetcher;
+import com.sbss.bithon.agent.rpc.brpc.setting.ISettingFetcher;
 import com.sbss.bithon.agent.rpc.brpc.event.IEventCollector;
 import com.sbss.bithon.agent.rpc.brpc.metrics.IMetricCollector;
 import com.sbss.bithon.agent.rpc.brpc.tracing.ITraceCollector;
